Discard metrics received too late
Under higher load, it may happen that we do not empty the pipe fast
enough. This leads to first the agent not responding to mailbox messages
in time and the malamute outgoing queue filling. To avoid this, fetch
all pending metric updates before doing any evaluation and if the same
metric is received more than once, only consider the latest version. We
may miss some spikes and sags in the metrics this way, but we already
miss these by only polling once per 30 seconds.